Move AddressSpace from searchlib/attribute -> searchlib/common.
