+++ searchlib/src/vespa/searchlib/attribute/CMakeLists.txt
@@ -1,7 +1,6 @@
 # Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchlib_attribute OBJECT
     SOURCES
-    address_space.cpp
     address_space_usage.cpp
     attribute.cpp
     attribute_blueprint_factory.cpp
