+++ searchlib/src/vespa/searchlib/common/address_space.cpp
