+++ searchlib/src/vespa/searchlib/common/address_space.h
